+++ apps/chat/src/types/error.ts
@@ -3,3 +3,17 @@ export interface ErrorMessage {
   title: string;
   messageLines: string[];
 }
+
+export class OpenAIError extends Error {
+  type: string;
+  param: string;
+  code: string;
+
+  constructor(message: string, type: string, param: string, code: string) {
+    super(message);
+    this.name = 'OpenAIError';
+    this.type = type;
+    this.param = param;
+    this.code = code;
+  }
+}
